+++ docs/changes.md
@@ -2,6 +2,7 @@
 
 2022/01/05
   * fscript公开函数value\_id/value\_func。
+  * fscript增加log10/pow10等函数。
 
 2022/01/04
   * fscript支持保存函数名，方便调试。